requirements.txt
# Combined requirements for TrafficCrash project
# Producer dependencies
kafka-python
pandas

# Consumer dependencies
pyspark
hdfs

